Finish the Voice Alerts and Rainbow Brackets renames
The earlier rename changed the directories and plugin.name but left the
package identity and every prose string alone. The Gradle job summary in
the Publish addons workflow reports rootProject.name, so it still listed
"Beepy" and "rainbow-on-the-go" after a publish.

Rename both add-ons end to end:

  com.example.beepy               -> org.appdevforall.voicealerts
  Beepy                           -> VoiceAlertsPlugin
  org.appdevforall.rainbowonthego -> org.appdevforall.rainbowbrackets
  RainbowOnTheGoPlugin            -> RainbowBracketsPlugin

This covers plugin.id, applicationId, namespace, the Kotlin package
directories, class names, rootProject.name, android:label, app_name,
proguard keep rules, both READMEs, both add-on HTML pages, and the
Rainbow Brackets Tier 3 help page and tooltip text.

Also fix getTooltipCategory() in Rainbow Brackets, which returned the
hand-written "plugin_rainbow_on_the_go" instead of "plugin_" plus the
plugin id. Per CLAUDE.md that mismatch makes the tooltip render "n/a".
It now derives from PLUGIN_ID.

Breaking: plugin.id changes, so the Plugin Manager treats these as new
add-ons. An existing install does not upgrade in place; it must be
removed and reinstalled.
