+++ plugins/Rainbow-Brackets/proguard-rules.pro
@@ -1,4 +1,4 @@
 # Add project specific ProGuard rules here.
 
 # Keep plugin classes — the IDE loads them by reflection via plugin.main_class.
--keep class org.appdevforall.rainbowonthego.** { *; }
+-keep class org.appdevforall.rainbowbrackets.** { *; }
