+++ plugins/Rainbow-Brackets/rainbow-brackets.html
@@ -68,7 +68,7 @@
 
 <section id="overview">
   <h2>1. Executive Overview</h2>
-  <p>RainbowOnTheGo makes nested code easier to read by coloring each pair of
+  <p>Rainbow Brackets makes nested code easier to read by coloring each pair of
   brackets according to how deeply it is nested. An opening delimiter and its matching
   closing delimiter share the same color, so you can see at a glance which <code>(</code>
   pairs with which <code>)</code>, even in deeply nested expressions.</p>
