+++ plugins/Rainbow-Brackets/src/main/AndroidManifest.xml
@@ -7,7 +7,7 @@
         <!-- Plugin metadata -->
         <meta-data
             android:name="plugin.id"
-            android:value="org.appdevforall.rainbowonthego" />
+            android:value="org.appdevforall.rainbowbrackets" />
 
         <meta-data
             android:name="plugin.name"
@@ -38,7 +38,7 @@
 
         <meta-data
             android:name="plugin.main_class"
-            android:value="org.appdevforall.rainbowonthego.RainbowOnTheGoPlugin" />
+            android:value="org.appdevforall.rainbowbrackets.RainbowBracketsPlugin" />
 
         <!-- Plugin Manager icons. Required for debug builds; release builds
              skip this check. Paths are inside the .cgp (APK) zip — assets/
