+++ plugins/Rainbow-Brackets/src/main/res/values/strings.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <string name="app_name">RainbowOnTheGo</string>
+    <string name="app_name">Rainbow Brackets</string>
 </resources>
