+++ plugins/Random-XKCD/build.gradle.kts
@@ -84,7 +84,7 @@ tasks.wrapper {
 }
 
 // Disable AAR metadata checks that fail under the plugin-builder
-// pipeline (Beepy + Forms use the same workaround).
+// pipeline (Voice Alerts + Forms use the same workaround).
 tasks.matching {
     it.name.contains("checkDebugAarMetadata") ||
     it.name.contains("checkReleaseAarMetadata")
