+++ plugins/Voice-Alerts/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <string name="app_name">Beepy</string>
+    <string name="app_name">Voice Alerts</string>
     <string name="plugin_description">Add earcons to build status</string>
 </resources>
