+++ plugins/Voice-Alerts/voice-alerts.html
@@ -91,7 +91,7 @@
 <header>
   <h1>Voice Alerts</h1>
   <p class="subtitle">Audible build-status earcons for Code On The Go</p>
-  <div class="meta">Version 1.0.0 &middot; Author: App Dev for All &middot; Package: <code>com.example.beepy</code></div>
+  <div class="meta">Version 1.0.0 &middot; Author: App Dev for All &middot; Package: <code>org.appdevforall.voicealerts</code></div>
 </header>
 
 <nav>
@@ -110,7 +110,7 @@
 <section id="overview">
   <h2>1. Executive Overview</h2>
   <p>
-    Beepy is a lightweight Code On The Go plugin that delivers audible feedback
+    Voice Alerts is a lightweight Code On The Go plugin that delivers audible feedback
     during builds. It plays a distinct short sound (earcon) when a build starts,
     succeeds, or fails, letting developers keep their eyes on code while staying
     aware of build progress.
@@ -175,7 +175,7 @@
         <li>src/main/
           <ul>
             <li>AndroidManifest.xml</li>
-            <li>kotlin/com/example/beepy/<strong>Beepy.kt</strong></li>
+            <li>kotlin/org/appdevforall/voicealerts/<strong>VoiceAlertsPlugin.kt</strong></li>
             <li>res/raw/ &mdash; started.wav, finished.wav, failed.wav</li>
             <li>res/values/strings.xml</li>
           </ul>
@@ -193,7 +193,7 @@
             +------------+-------------+
                          |
     +--------------------------------------+
-    |               Beepy                  |
+    |         VoiceAlertsPlugin            |
     +--------------------------------------+
     | - context : PluginContext            |
     | - soundPool : SoundPool?             |
@@ -287,19 +287,19 @@
 <section id="integration">
   <h2>4. Integration Points</h2>
   <p>
-    Beepy touches the host IDE through a small number of well-defined
+    Voice Alerts touches the host IDE through a small number of well-defined
     interfaces. It does not modify any main-app source files &mdash; all
     integration happens at runtime through the plugin framework.
   </p>
 
   <h3>4.1 Plugin framework (<code>IPlugin</code>)</h3>
   <p>
-    The <code>Beepy</code> class implements
+    The <code>VoiceAlertsPlugin</code> class implements
     <code>com.itsaky.androidide.plugins.IPlugin</code>, which the IDE
     discovers and manages. The manifest declares the entry point:
   </p>
   <pre><code>&lt;meta-data android:name="plugin.main_class"
-           android:value="com.example.beepy.Beepy" /&gt;</code></pre>
+           android:value="org.appdevforall.voicealerts.VoiceAlertsPlugin" /&gt;</code></pre>
   <p>
     The IDE calls <code>initialize()</code> &rarr; <code>activate()</code>
     on load, and <code>deactivate()</code> &rarr; <code>dispose()</code> on
@@ -330,7 +330,7 @@
 
   <h3>4.4 Permissions</h3>
   <p>
-    Beepy requires no plugin permissions. The available permission system
+    Voice Alerts requires no plugin permissions. The available permission system
     (<code>filesystem.read</code>, <code>filesystem.write</code>,
     <code>network.access</code>, <code>system.commands</code>,
     <code>ide.settings</code>, <code>project.structure</code>) is left
@@ -354,7 +354,7 @@
            |           |
            v           v
        +------------------------+
-       |         Beepy          |
+       |   VoiceAlertsPlugin    |
        +------------------------+
        | onBuildStarted()       |-----&gt; play("started")
        | onBuildFinished()      |-----&gt; play("finished")
@@ -380,7 +380,7 @@
   <ol>
     <li>Open Code On The Go's plugin manager.</li>
     <li>Import the <code>.cgp</code> file.</li>
-    <li>The IDE discovers <code>Beepy</code> via the manifest metadata and
+    <li>The IDE discovers <code>VoiceAlertsPlugin</code> via the manifest metadata and
         activates it automatically.</li>
   </ol>
 
@@ -432,7 +432,7 @@
 </section>
 
 <footer>
-  Beepy Plugin Documentation &middot; Version 1.0.0 &middot; com.example.beepy
+  Voice Alerts Plugin Documentation &middot; Version 1.0.0 &middot; org.appdevforall.voicealerts
 </footer>
 
 </body>
